terraform/cloud_function/requirements.txt
google-api-python-client
google-cloud-bigquery
